выполнил 4 задание добавил модуляцию
